[FIX] GradeType과 semesterType을 constance로 옮김
