+++ src/constance/grade/index.ts
@@ -6,7 +6,7 @@ export const GRADE = [
   { grade: 'C', id: 'c' },
   { grade: 'D', id: 'd' },
   { grade: 'E', id: 'e' },
-  { grade: '0', id: '0' },
+  { grade: 'X', id: 'X' },
 ];
 export const GRADEANDSEMESTER = [
   { id: 'freshmanFirst', grade: 1, semester: 1 },
@@ -14,5 +14,32 @@ export const GRADEANDSEMESTER = [
   { id: 'sophomoreFirst', grade: 2, semester: 1 },
   { id: 'sophomoreSecond', grade: 2, semester: 2 },
   { id: 'seniorFirst', grade: 3, semester: 1 },
+  { id: 'seniorSecond', grade: 3, semester: 2 },
 ];
 export const SUBJECT = ['korean', 'social', 'history', 'math', 'science', 'technical', 'english'];
+
+export type GradeType = {
+  korean: string;
+  social: string;
+  history: string;
+  math: string;
+  science: string;
+  technical: string;
+  english: string;
+  isCheck: {
+    freshmanFirst: boolean;
+    freshmanSecond: boolean;
+    sophomoreFirst: boolean;
+    sophomoreSecond: boolean;
+    seniorFirst: boolean;
+    seniorSecond: boolean;
+  };
+};
+export type semesterType = {
+  freshmanFirst: boolean;
+  freshmanSecond: boolean;
+  sophomoreFirst: boolean;
+  sophomoreSecond: boolean;
+  seniorFirst: boolean;
+  seniorSecond: boolean;
+};
